Step 5b: CallsTable basically compelte
